backend/requirements.txt
django==5.2.10
djangorestframework==3.16.1
django-cors-headers==4.9.0
PyMySQL==1.1.1

